修复decimal和 浮点数区分，datetime 也属于timestamp

+++ devs/src/main/java/com/delmar/devs/ftl/ModelGenerateStrategy.java
@@ -54,7 +54,7 @@ public class ModelGenerateStrategy {
                     {
                         param.put("hasDate",true);
                     }
-                    if(columnInfo.getDataType()==ColumnDataType.FLOAT.getKey())
+                    if(columnInfo.getDataType()==ColumnDataType.DECIMAL.getKey())
                     {
                         param.put("hasDecimal",true);
                     }
